requirements.txt
Flask
requests
python-dotenv
gunicorn
setuptools
wheel